requirements.txt
joblib==0.10.2
kafka-python==1.3.1
